+++ src/pages/common/NameSettingPage/index.tsx
